+++ src/calculators/gt_calculator.h
@@ -10,21 +10,14 @@
 #ifndef GTCALCULATOR_H
 #define GTCALCULATOR_H
 
-#include "gt_calculators_exports.h"
-
 #include "gt_processcomponent.h"
 
 #include <QPointer>
-
-#include "gt_boolproperty.h"
-#include "gt_objectpath.h"
 
 class QDir;
 class GtModeProperty;
 class GtLabelProperty;
 class GtAbstractRunnable;
-class GtObjectLinkProperty;
-class GtObjectPathProperty;
 
 /**
  * @brief The GtCalculator class
@@ -43,7 +36,7 @@ public:
      * @brief Execution process of the calculator.
      * @return Whether execution process was successful or not.
      */
-    bool exec();
+    bool exec() Q_DECL_OVERRIDE;
 
     /**
      * @brief Main run method of the calculator.
@@ -93,29 +86,11 @@ public:
     void setExecutionLabel(const QString& label);
 
     /**
-     * @brief Returns true if calculator is set to skip mode.
-     * @return Skip indicator.
-     */
-    bool isSkipped();
-
-    /**
-     * @brief Sets calculator skip mode.
-     * @param Whether calculator should be skipped or not.
-     */
-    void setSkipped(bool val);
-
-    /**
      * @brief Returns true if run should be marked as failed if warning flag
      * is activated. Otherwise false is returnes.
      * @return Whether run should be marked as failed or not.
      */
     bool runFailsOnWarning();
-
-    /**
-     * @brief Sets current state of process component.
-     * @param New state.
-     */
-    void setState(GtProcessComponent::STATE state) Q_DECL_OVERRIDE;
 
     /**
      * @brief Returns datamodel object based on given object uuid. If no
@@ -182,35 +157,6 @@ public:
         return Q_NULLPTR;
     }
 
-    /**
-     * @brief Returns datamodel object based on given object link property.
-     * If no object is found nullpointer is returned.
-     * @tparam T Object type
-     * @param prop Object link property
-     * @return Object corresponding to given object link property
-     */
-    template <class T>
-    T data(GtObjectLinkProperty& prop)
-    {
-        const QString uuid = dataHelper(prop);
-        return data<T>(uuid);
-    }
-
-    /**
-     * @brief Returns datamodel object based on given object path property.
-     * If no object is found nullpointer is returned.
-     * @tparam T Object type
-     * @param prop Object path property
-     * @return Object corresponding to given object path property
-     */
-    template <class T>
-    T data(GtObjectPathProperty& prop)
-    {
-        const GtObjectPath path = pathHelper(prop);
-
-        return data<T>(path);
-    }
-
 protected:
     /**
      * @brief Constructor
@@ -258,25 +204,8 @@ private:
     /// Pointer to runnable associated to the calculator.
     GtAbstractRunnable* m_runnable;
 
-    /// Skip indicator.
-    GtBoolProperty m_skipped;
-
     /// Fail run on warning indicator.
     GtBoolProperty m_failRunOnWarning;
-
-    /**
-     * @brief Returns uuid string of given object link property.
-     * @param prop Object link property.
-     * @return Uuid string.
-     */
-    QString dataHelper(GtObjectLinkProperty& prop);
-
-    /**
-     * @brief Returns object path of given object path property.
-     * @param prop Object path property.
-     * @return Object path.
-     */
-    GtObjectPath pathHelper(GtObjectPathProperty& prop);
 
 };