+++ CHANGELOG.md
@@ -15,12 +15,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Saving a project was moved to a separate thread - #127
 - Changed the class structure for Datazones to make use of the externalization feature - #127
+- Sub-Tasks can be skipped - #82
 
 ### Added
 - Multiple labels can be selected and deleted at once - #111
 - Integrated the hdf5 data format and added support to ex-/internalize datasets at runtime including the option to enable this feature on a project basis - #127
 - Shortcut settings for GTlab including the introduction of several new shortcuts and the option to maintain them in the preferences - #112
-- Versionnumber based on GTVersionnumber for tasks - #129
+- Versionnumber based on GtVersionNumber for tasks - #129
 
 ### Removed
 - Nozzle map was removed to be added in the map collection - #114